+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- First element selection on `SCombobox` component.
+- Default symbol on amount input
+- Autofill style on SInput
 
 ## 2.0.0-beta.31 - 2023-11-01
 ### Changed